Update Netty response headers when received
Prior to this commit, the NettyHttpChannel would update its
`nettyResponse` attribute when it reads the response message from the
netty channel. But this instance also has a `responseHeaders` attribute
which is derived from the netty response itself. Whenever the new
response is received, those headers would not be updated, thus still
pointing to the initial values set in the constructor.

This commit updates the `responseHeaders` attribute when the response
message is received on the netty channel.